# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "B2B Marketplace API",
    "description": "API for B2B Marketplace Platform",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server (HTTP)"
    },
    {
      "url": "https://localhost:5001",
      "description": "Development server (HTTPS)"
    }
  ],
  "paths": {
...
    "/api/admin/certifications": {
      "get": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.PagedResultDto`1[[B2BMarketplace.Core.DTOs.CertificationDto, B2BMarketplace.Core, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.PagedResultDto`1[[B2BMarketplace.Core.DTOs.CertificationDto, B2BMarketplace.Core, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.PagedResultDto`1[[B2BMarketplace.Core.DTOs.CertificationDto, B2BMarketplace.Core, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminCertification"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/certifications/{id}": {
      "get": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.CertificationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/admin/certifications/search": {
      "get": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "searchTerm",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.PagedResultDto`1[[B2BMarketplace.Core.DTOs.CertificationDto, B2BMarketplace.Core, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.PagedResultDto`1[[B2BMarketplace.Core.DTOs.CertificationDto, B2BMarketplace.Core, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/B2BMarketplace.Core.DTOs.PagedResultDto`1[[B2BMarketplace.Core.DTOs.CertificationDto, B2BMarketplace.Core, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/certifications/{id}/approve": {
      "post": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/admin/certifications/{id}/reject": {
      "post": {
        "tags": [
          "AdminCertification"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Api.Controllers.Admin.RejectCertificationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Api.Controllers.Admin.RejectCertificationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/B2BMarketplace.Api.Controllers.Admin.RejectCertificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Certifications/{id}": {
      "get": {
        "tags": [
          "Certifications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "B2BMarketplace.Api.Controllers.Admin.RejectCertificationRequest": {
        "type": "object",
        "properties": {
          "RejectionReason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "B2BMarketplace.Core.DTOs.CertificationDto": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "Name": {
            "type": "string",
            "nullable": true
          },
          "Status": {
            "$ref": "#/components/schemas/B2BMarketplace.Core.Enums.CertificationStatus"
          },
          "SubmittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "ReviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "AdminNotes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "B2BMarketplace.Core.Enums.CertificationStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}